Renamed files and cleaned code
