+++ pkg/deepinference/ffi.go
@@ -5,13 +5,13 @@
 
 //go:build (linux || darwin) && cgo
 
-package get_text_embeddings
+package deepinference
 
 /*
-#cgo CFLAGS: -I${SRCDIR}/../../../../pkg/get_text_embeddings/rust/include
-#cgo LDFLAGS: -L${SRCDIR}/../../../../pkg/get_text_embeddings/rust/target/release -lget_text_embeddings -Wl,-rpath,${SRCDIR}/../../../../pkg/get_text_embeddings/rust/target/release
+#cgo CFLAGS: -I${SRCDIR}/rust/include
+#cgo LDFLAGS: -L${SRCDIR}/rust/target/release -ldeepinference -Wl,-rpath,${SRCDIR}/rust/target/release
 #include <stdlib.h>
-#include "get_text_embeddings.h"
+#include "deepinference.h"
 */
 import "C"
 
@@ -22,15 +22,15 @@ import (
 
 func Init() error {
 	var err *C.char
-	C.dd_get_text_embeddings_init(&err)
+	C.dd_deepinference_init(&err)
 	if err != nil {
-		return fmt.Errorf("failed to initialize get_text_embeddings: %s", C.GoString(err))
+		return fmt.Errorf("failed to initialize deepinference: %s", C.GoString(err))
 	}
 	return nil
 }
 
 func GetEmbeddingsSize() (int, error) {
-	size := C.dd_get_text_embeddings_get_embeddings_size()
+	size := C.dd_deepinference_get_embeddings_size()
 	return int(size), nil
 }
 
@@ -45,7 +45,7 @@ func GetEmbeddings(text string) ([]float32, error) {
 
 	buffer := make([]float32, size)
 	var errPtr *C.char
-	C.dd_get_text_embeddings_get_embeddings(cText, (*C.float)(unsafe.Pointer(&buffer[0])), &errPtr)
+	C.dd_deepinference_get_embeddings(cText, (*C.float)(unsafe.Pointer(&buffer[0])), &errPtr)
 	if errPtr != nil {
 		return nil, fmt.Errorf("failed to get embeddings: %s", C.GoString(errPtr))
 	}
